move fixtures for testing qc_utils into conftest.py
